Allow border radius if only one control in field
If there is only one child inside a field with addons, the current first and last child selectors remove any radius. Since a single control would be both a first and last child at the same time, it gets canceled out. This leverages the :only-child property to not apply this rule if there is only one control.